+++ libraries/datetime/src/Timestamp.ts
@@ -4,11 +4,11 @@ import { TimeUnit } from './TimeUnit.ts';
  * Represents a point in time.
  */
 export class Timestamp {
-  _milliseconds: number;
+  _millis: number;
   _timeUnit: TimeUnit;
 
   // region | Static methods
-  static fromMilliseconds(milliseconds: number): Timestamp {
+  static fromMillis(milliseconds: number): Timestamp {
     return new Timestamp(new Date(milliseconds));
   }
 
@@ -19,30 +19,23 @@ export class Timestamp {
 
   // region | Public methods
   constructor(dateTime?: TimestampInput, options: { timeUnit?: TimeUnit } = {}) {
-    const { timeUnit = TimeUnit.Milliseconds } = options;
+    const { timeUnit = TimeUnit.Millis } = options;
     if (dateTime instanceof Timestamp) {
-      this._milliseconds = dateTime._milliseconds;
+      this._millis = dateTime._millis;
     } else if (dateTime) {
-      this._milliseconds = new Date(dateTime).getTime();
+      this._millis = new Date(dateTime).getTime();
     } else {
-      this._milliseconds = Date.now();
+      this._millis = Date.now();
     }
     this._timeUnit = timeUnit;
   }
 
-  get milliseconds(): number {
-    return this._milliseconds;
-  }
-
-  /**
-   * Alias for `milliseconds`.
-   */
-  get ms(): number {
-    return this._milliseconds;
+  get millis(): number {
+    return this._millis;
   }
 
   get seconds(): number {
-    return TimeUnit.convert(this._milliseconds, TimeUnit.Milliseconds, TimeUnit.Seconds);
+    return TimeUnit.convert(this._millis, TimeUnit.Millis, TimeUnit.Seconds);
   }
 
   get timeUnit(): TimeUnit {
@@ -50,7 +43,7 @@ export class Timestamp {
   }
 
   clone(options: { timeUnit?: TimeUnit } = {}): Timestamp {
-    return new Timestamp(this._milliseconds, options);
+    return new Timestamp(this._millis, options);
   }
 
   public setTimeUnit(timeUnit: TimeUnit): Timestamp {
@@ -59,19 +52,39 @@ export class Timestamp {
   }
 
   toDate(): Date {
-    return new Date(this._milliseconds);
+    return new Date(this._millis);
   }
 
-  toLocaleDateTimeString(locales?: Intl.LocalesArgument, options?: Intl.DateTimeFormatOptions): string {
-    const date = new Date(this._milliseconds);
-    return date.toLocaleDateString(locales, options) + ' ' + date.toLocaleTimeString(locales, options);
+  /**
+   * Returns a timestamp in the format `YYYY-MM-DD HH:MM[:SS[.000]] UTC` (depending on the granularity).
+   * It is the same format as `Date.toISOString()` but with a space instead of a `T` and ` UTC` instead of `Z`.
+   * This date format is human-readable, sortable, and accepted by the Date constructor.
+   * TODO: Decide how to handle hours.
+   */
+  toHumaneUtcString(options: TimestampFormatOptions = {}): string {
+    const { granularity = this.timeUnit } = options;
+
+    if (granularity === TimeUnit.Hours) {
+      throw new Error('Method does not support TimeUnit.Hours granularity.');
+    }
+
+    const isoString = this.toIsoString({ granularity });
+    if (granularity.inMillis <= TimeUnit.Minutes.inMillis) {
+      return isoString
+        .replace(/T/, ' ')
+        .replace(/Z$/, ' UTC');
+    }
+    return `${isoString} UTC`;
   }
 
-  toIsoString(options: { granularity?: TimeUnit } = {}): string {
+  /**
+   * Returns a timestamp in the format `YYYY-MM-DDTHH:MM[:SS[.000]]Z` (depending on the granularity).
+   */
+  toIsoString(options: TimestampFormatOptions = {}): string {
     const { granularity = this.timeUnit } = options;
     const isoDateTime = this.toDate().toISOString();
     switch (granularity) {
-      case TimeUnit.Milliseconds:
+      case TimeUnit.Millis:
         return isoDateTime;
       case TimeUnit.Seconds:
         return isoDateTime.replace(/\.\d{3}Z$/, 'Z');
@@ -86,6 +99,11 @@ export class Timestamp {
     }
   }
 
+  toLocaleDateTimeString(locales?: Intl.LocalesArgument, options?: Intl.DateTimeFormatOptions): string {
+    const date = new Date(this._millis);
+    return date.toLocaleDateString(locales, options) + ' ' + date.toLocaleTimeString(locales, options);
+  }
+
   toString(): string {
     return this.toIsoString();
   }
@@ -93,4 +111,8 @@ export class Timestamp {
 
 // region --- Types ---
 export type TimestampInput = Date | number | string | Timestamp;
+
+interface TimestampFormatOptions {
+  granularity?: TimeUnit | undefined;
+}
 // endregion - Types